+++ origin/src/Model/Behavior/Behavior.php
@@ -17,7 +17,6 @@ namespace Origin\Model\Behavior;
 use Origin\Model\Model;
 use Origin\Model\Entity;
 use Origin\Core\ConfigTrait;
-use Origin\Core\Logger;
 
 class Behavior
 {
@@ -147,17 +146,5 @@ class Behavior
     public function model()
     {
         return $this->_model;
-    }
-
-
-    /**
-     * Returns a Logger Object
-     *
-     * @param string $channel
-     * @return \Origin\Core\Logger
-     */
-    public function logger(string $channel = 'Behavior')
-    {
-        return new Logger($channel);
     }
 }